fixed everyt

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/vars/states.java
@@ -3,6 +3,6 @@ package org.firstinspires.ftc.teamcode.vars;
 import com.pedropathing.geometry.Pose;
 
 public class states {
-    public static float turretPos = 0F;
+    public static double turretPos = 0F;
     public static Pose autoEndPose = new Pose(58.000, 104.000, Math.toRadians(147));
 }
